envUnity(feat): lump previous states into a larger state
envUnity has a new class Env1D. This is specifically tied to
environments whose states are 1D in nature. This state will
additionally be able to generate a state that concats the
past N states together into a single state. Being able to
retain the past information is critical to being able to
perforom well in most games.